+++ app/Models/DraftOrderItem.php
@@ -60,6 +60,6 @@ class DraftOrderItem extends Model
      */
     public function menuItem(): BelongsTo
     {
-        return $this->belongsTo(MenuItem::class);
+        return $this->belongsTo(MenuItem::class)->withTrashed();
     }
 }