+++ app/Models/KitchenTicketItem.php
@@ -66,7 +66,7 @@ class KitchenTicketItem extends Model
      */
     public function menuItem(): BelongsTo
     {
-        return $this->belongsTo(MenuItem::class);
+        return $this->belongsTo(MenuItem::class)->withTrashed();
     }
 
     /**