+++ app/Models/OrderItem.php
@@ -61,7 +61,7 @@ class OrderItem extends Model
      */
     public function menuItem(): BelongsTo
     {
-        return $this->belongsTo(MenuItem::class);
+        return $this->belongsTo(MenuItem::class)->withTrashed();
     }
 
     /**